Extract AgentCardDetails from agent-card.tsx (#tech-debt)
Split the expanded detail panel (git info, feedback panel, action buttons)
into agent-card-details.tsx, reducing agent-card.tsx from 758 to 458 lines.
Moved CompactMetaRow helper and useAgentDiffStats/useCopyText hooks into the
new component since they're only used in the expanded view.
