+++ apps/web/src/components/app/agent-card-details.tsx
@@ -1,0 +1,380 @@
+import React from "react";
+import {
+  AlertTriangle,
+  Archive,
+  Check,
+  Copy,
+  Folder,
+  FolderTree,
+  GitBranch,
+  Pause,
+  Play,
+} from "lucide-react";
+
+import { AgentMeta, FrontTruncatedValue } from "@/components/app/agent-meta";
+import { DiffStatBadge } from "@/components/app/diff-stat-badge";
+import { useAgentDiffStats } from "@/hooks/use-agent-diff-stats";
+import {
+  type FeedbackDetailState,
+  ParentFeedbackPanel,
+} from "@/components/app/feedback-panel";
+import { IdeLaunchButton } from "@/components/app/ide-launch-button";
+import { PersonaLauncher } from "@/components/app/persona-launcher";
+import { type Agent, type AgentVisualState } from "@/components/app/types";
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+
+import { useCopyText } from "@/hooks/use-copy";
+import { type AgentType } from "@/lib/agent-types";
+import { type IdeType } from "@/lib/ide-types";
+import { cn } from "@/lib/utils";
+
+function CompactMetaRow({
+  label,
+  icon,
+  value,
+  mono = false,
+  truncateStart = false,
+}: {
+  label: string;
+  icon?: React.ReactNode;
+  value: string;
+  mono?: boolean;
+  truncateStart?: boolean;
+}): JSX.Element {
+  return (
+    <div className="flex items-start justify-between gap-3">
+      <span className="shrink-0 text-muted-foreground/75" title={label}>
+        {icon ?? (
+          <span className="text-[10px] uppercase tracking-wide">{label}</span>
+        )}
+      </span>
+      <div className="min-w-0 flex-1 text-left">
+        {truncateStart ? (
+          <FrontTruncatedValue
+            value={value}
+            mono={mono}
+            className="text-left text-foreground"
+            tooltipClassName="max-w-[420px]"
+          />
+        ) : (
+          <div
+            className={cn(
+              "text-foreground break-all",
+              mono && "font-mono text-[11px]"
+            )}
+          >
+            {value}
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
+
+export type AgentCardDetailsProps = {
+  agent: Agent;
+  agents: Agent[];
+  childAgents: Agent[];
+  selectedAgentId: string | null;
+  agentVisualState: (agent: Agent) => AgentVisualState;
+  isFullAccessEnabled: (
+    agent: Pick<Agent, "agentArgs" | "fullAccess">
+  ) => boolean;
+  detachTerminal: () => void;
+  attachToAgent: (agent: Agent) => Promise<void>;
+  startAgent: (agent: Agent) => Promise<void>;
+  setDeleteTarget: (agent: Agent | null) => void;
+  setDeleteConfirmOpen: (open: boolean) => void;
+  setStopTarget: (agent: Agent | null) => void;
+  setStopConfirmOpen: (open: boolean) => void;
+  sendTerminalInput?: (data: string) => void;
+  connectedAgentId?: string | null;
+  onOpenFeedbackDetail?: (state: FeedbackDetailState) => void;
+  feedbackDetailState?: FeedbackDetailState;
+  onRequestClose?: () => void;
+  closeOnSessionAction?: boolean;
+  enabledAgentTypes: AgentType[];
+  enabledIdes: IdeType[];
+};
+
+export function AgentCardDetails({
+  agent,
+  agents,
+  childAgents,
+  selectedAgentId,
+  agentVisualState: getVisualState,
+  isFullAccessEnabled,
+  detachTerminal,
+  attachToAgent,
+  startAgent,
+  setDeleteTarget,
+  setDeleteConfirmOpen,
+  setStopTarget,
+  setStopConfirmOpen,
+  sendTerminalInput,
+  connectedAgentId,
+  onOpenFeedbackDetail,
+  feedbackDetailState,
+  onRequestClose,
+  closeOnSessionAction = false,
+  enabledAgentTypes,
+  enabledIdes,
+}: AgentCardDetailsProps): JSX.Element {
+  const state = getVisualState(agent);
+  const isStopped = state === "stopped";
+  const isTerminalAgent = agent.type === "terminal";
+  const fullAccessEnabled = isFullAccessEnabled(agent);
+  const sidebarBaseBranch = agent.baseBranch ?? "main";
+  const [worktreePathCopied, copyWorktreePath] = useCopyText();
+  const { diffStats, refresh: refreshDiffStats } = useAgentDiffStats(
+    agent.id,
+    true
+  );
+
+  return (
+    <>
+      <div className="flex flex-col gap-3 px-0 pb-0.5 pt-0.5">
+        <div className="space-y-1.5">
+          <div className="relative space-y-2 rounded-xl border border-border/60 bg-background/25 px-3 py-3 text-xs text-muted-foreground">
+            <div className="absolute right-3 top-3">
+              <DiffStatBadge
+                diffStats={diffStats}
+                latestEventAt={agent.latestEvent?.updatedAt ?? null}
+                onRefresh={refreshDiffStats}
+              />
+            </div>
+            {agent.gitContext?.isWorktree ? (
+              <>
+                <div className="flex items-start justify-between gap-3">
+                  <span
+                    className="shrink-0 text-muted-foreground/75"
+                    title="Branch"
+                  >
+                    <GitBranch className="h-3.5 w-3.5" />
+                  </span>
+                  <div className="min-w-0 flex-1 text-left">
+                    <div className="text-muted-foreground">
+                      <FrontTruncatedValue
+                        value={sidebarBaseBranch}
+                        mono
+                        className="text-left text-muted-foreground"
+                        tooltipClassName=""
+                        tooltipValue={`Base branch: ${sidebarBaseBranch}`}
+                      />
+                    </div>
+                    <div className="mt-0.5 flex items-center gap-1 pl-2">
+                      <span className="shrink-0 font-mono text-[11px] text-muted-foreground/50">
+                        ↳
+                      </span>
+                      <FrontTruncatedValue
+                        value={agent.gitContext.branch}
+                        mono
+                        className="text-left"
+                        tooltipValue={`Working branch: ${agent.gitContext.branch}`}
+                      />
+                    </div>
+                  </div>
+                </div>
+              </>
+            ) : (
+              <>
+                <CompactMetaRow
+                  label="Working dir"
+                  icon={<Folder className="h-3.5 w-3.5" />}
+                  value={agent.cwd}
+                  mono
+                  truncateStart
+                />
+                {agent.gitContext ? (
+                  <CompactMetaRow
+                    label="Branch"
+                    icon={<GitBranch className="h-3.5 w-3.5" />}
+                    value={agent.gitContext.branch}
+                    mono
+                    truncateStart
+                  />
+                ) : null}
+              </>
+            )}
+            <div className="flex items-center justify-between gap-2 pt-1">
+              <div className="flex items-center gap-2">
+                {agent.cwd ? (
+                  <IdeLaunchButton path={agent.cwd} enabledIdes={enabledIdes} />
+                ) : null}
+                {agent.gitContext?.isWorktree && agent.cwd ? (
+                  <Tooltip>
+                    <TooltipTrigger asChild>
+                      <Button
+                        variant="ghost"
+                        size="sm"
+                        onClick={() => agent.cwd && copyWorktreePath(agent.cwd)}
+                        aria-label={
+                          worktreePathCopied
+                            ? "Worktree path copied"
+                            : `Copy worktree path: ${agent.cwd}`
+                        }
+                        className="group relative h-auto min-h-6 gap-1 rounded-full border border-border bg-muted/35 px-2 py-0.5 text-[10px] font-normal text-muted-foreground before:absolute before:inset-x-0 before:-inset-y-1.5 before:content-[''] hover:bg-muted/60 hover:text-foreground"
+                      >
+                        {worktreePathCopied ? (
+                          <Check className="h-3 w-3 text-status-done" />
+                        ) : (
+                          <>
+                            <FolderTree className="h-3 w-3 group-hover:hidden group-focus-visible:hidden" />
+                            <Copy className="hidden h-3 w-3 group-hover:block group-focus-visible:block" />
+                          </>
+                        )}
+                        <span>Worktree</span>
+                        <span className="sr-only" aria-live="polite">
+                          {worktreePathCopied ? "Worktree path copied" : ""}
+                        </span>
+                      </Button>
+                    </TooltipTrigger>
+                    <TooltipContent className="max-w-[420px] break-all">
+                      {agent.cwd}
+                      <div className="mt-1 text-[10px] opacity-70">
+                        Click to copy
+                      </div>
+                    </TooltipContent>
+                  </Tooltip>
+                ) : null}
+              </div>
+              {isTerminalAgent ? (
+                <span />
+              ) : (
+                <div
+                  className={cn(
+                    "inline-flex min-h-6 items-center gap-1 rounded-full px-2 py-0.5 text-[10px]",
+                    fullAccessEnabled
+                      ? "border border-status-waiting/35 bg-status-waiting/10 text-status-waiting"
+                      : "border border-border bg-muted/40 text-muted-foreground"
+                  )}
+                >
+                  {fullAccessEnabled ? (
+                    <AlertTriangle className="h-3 w-3" />
+                  ) : null}
+                  <span>{fullAccessEnabled ? "Full access" : "Sandboxed"}</span>
+                </div>
+              )}
+            </div>
+          </div>
+        </div>
+
+        {agent.lastError ? (
+          <AgentMeta label="Last error" value={agent.lastError} />
+        ) : null}
+        {agent.persona ? (
+          <div className="flex items-center gap-1.5">
+            <span className="uppercase tracking-wide text-[10px] text-muted-foreground/80">
+              Persona
+            </span>
+            <Badge variant="running">{agent.persona}</Badge>
+            {agent.parentAgentId ? (
+              <span className="text-[10px] text-muted-foreground">
+                from{" "}
+                {agents.find((a) => a.id === agent.parentAgentId)?.name ??
+                  agent.parentAgentId.slice(-6)}
+              </span>
+            ) : null}
+          </div>
+        ) : null}
+      </div>
+      {!agent.persona ? (
+        <div className="flex flex-col gap-3 px-0 pb-1">
+          {isTerminalAgent ? null : (
+            <ParentFeedbackPanel
+              parentAgentId={agent.id}
+              sendTerminalInput={sendTerminalInput}
+              isConnected={connectedAgentId === agent.id}
+              onRequestClose={onRequestClose}
+              closeOnSessionAction={closeOnSessionAction}
+              onOpenDetail={onOpenFeedbackDetail}
+              activeDetailItemId={
+                feedbackDetailState?.parentAgentId === agent.id &&
+                "itemId" in feedbackDetailState
+                  ? feedbackDetailState.itemId
+                  : null
+              }
+              childAgents={childAgents}
+              selectedAgentId={selectedAgentId}
+              agentVisualState={getVisualState}
+              detachTerminal={detachTerminal}
+              attachToAgent={attachToAgent}
+            />
+          )}
+          <div className="flex min-h-9 items-center justify-between gap-2 pt-2">
+            <div className="min-h-9 min-w-0 flex items-center">
+              {isTerminalAgent ? null : (
+                <PersonaLauncher
+                  agent={agent}
+                  enabledAgentTypes={enabledAgentTypes}
+                  disabled={isStopped || agent.status === "archiving"}
+                  disabledReason={
+                    isStopped
+                      ? "Agent is stopped — start it before launching a review."
+                      : agent.status === "archiving"
+                        ? "Agent is archiving."
+                        : undefined
+                  }
+                />
+              )}
+            </div>
+            <div className="flex h-9 shrink-0 items-center gap-[18px]">
+              {isStopped && agent.status !== "archiving" ? (
+                <Button
+                  size="sm"
+                  variant="ghost-primary"
+                  className="h-8 w-8 rounded-full border border-status-done/35 bg-status-done/10 p-0"
+                  aria-label="Resume"
+                  title="Resume"
+                  onClick={() => {
+                    if (closeOnSessionAction) onRequestClose?.();
+                    void startAgent(agent);
+                  }}
+                >
+                  <Play className="h-3.5 w-3.5" />
+                </Button>
+              ) : !isStopped && agent.status !== "archiving" ? (
+                <Button
+                  size="sm"
+                  variant="ghost-warning"
+                  className="h-8 w-8 rounded-full border border-status-waiting/35 bg-status-waiting/10 p-0"
+                  aria-label="Pause"
+                  title="Pause"
+                  onClick={() => {
+                    setStopTarget(agent);
+                    setStopConfirmOpen(true);
+                  }}
+                >
+                  <Pause className="h-3.5 w-3.5" />
+                </Button>
+              ) : null}
+              <Button
+                size="sm"
+                variant="ghost-destructive"
+                className="h-8 w-8 rounded-full border border-status-blocked/35 bg-status-blocked/10 p-0"
+                data-testid={`agent-archive-${agent.id}`}
+                aria-label="Archive"
+                title="Archive"
+                disabled={
+                  agent.status === "archiving" || agent.status === "creating"
+                }
+                onClick={() => {
+                  setDeleteTarget(agent);
+                  setDeleteConfirmOpen(true);
+                }}
+              >
+                <Archive className="h-3.5 w-3.5" />
+              </Button>
+            </div>
+          </div>
+        </div>
+      ) : null}
+    </>
+  );
+}
